+++ CHANGELOG.md
@@ -1,7 +1,7 @@
 ## [UNRELEASED] neptune 1.8.0
 
 ### Features
-- Added support for no synchronization callbacks ([#1478](https://github.com/neptune-ai/neptune-client/pull/1478))
+- Added support for callbacks that stop the synchronization if the lag or lack of progress exceeds a certain threshold ([#1478](https://github.com/neptune-ai/neptune-client/pull/1478))
 
 ### Fixes
 - Add newline at the end of generated `.patch` while tracking uncommitted changes ([#1473](https://github.com/neptune-ai/neptune-client/pull/1473))